pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taco"
version = "0.1.0"
description = "Pose graph optimization using IMU and CVGL image localization"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["pose-graph", "slam", "imu", "localization", "computer-vision", "sensor-fusion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "gtsam>=4.2.0",
    "networkx>=3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
cvgl = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "pillow>=10.0.0",
]
kitti = [
    "pykitti>=0.3.1",
    "pypose>=0.6.0",
    "osmnx>=1.6.0",
    "opencv-python>=4.8.0",
    "geopy>=2.4.0",
    "haversine>=2.8.0",
    "requests>=2.31.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "ruff>=0.8.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=1.25.0",
]
all = [
    "taco[cvgl,kitti,dev,docs]",
]

[project.urls]
Homepage = "https://github.com/yourusername/taco"
Documentation = "https://taco.readthedocs.io"
Repository = "https://github.com/yourusername/taco"
Issues = "https://github.com/yourusername/taco/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
taco = ["py.typed"]

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=taco",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
parallel = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EXE",    # flake8-executable
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",    # line too long (handled by black)
    "N803",    # allow uppercase argument names (T for transform matrix)
    "N806",    # allow uppercase variable names (T for transform, R for rotation)
    "N812",    # allow lowercase imports as non-lowercase (L for Lightning, F for functional)
    "PLR0913", # too many arguments
    "PLR0915", # too many statements
    "PLR2004", # magic values (use sparingly, prefer constants when reasonable)
    "SIM108",  # allow if-else blocks instead of ternary for clarity
    "TID252",  # allow relative imports from parent modules
    "RET504",  # allow assignment before return for clarity
    "ARG001",  # unused function arguments (kept for API consistency)
    "ARG002",  # unused method arguments (kept for API consistency)
    "ERA001",  # allow commented code (useful for documentation)
    "F401",    # allow unused imports in __init__ (re-exports)
    "PTH",     # allow os.path and open() (not always necessary to use pathlib)
    "UP006",   # allow List/Dict/Tuple in annotations (more explicit than list/dict/tuple)
    "UP035",   # allow typing.List/Dict/Tuple imports (more explicit than built-in generics)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # assert allowed in tests
    "ARG",    # unused arguments allowed in tests (fixtures)
    "PLR2004", # magic values allowed in tests
]
"examples/**/*.py" = [
    "PLR0915", # too many statements allowed in examples
    "PLR2004", # magic values allowed in examples
]
"docs/**/*.py" = [
    "PTH",    # allow os.path in docs
]

[tool.ruff.lint.isort]
known-first-party = ["taco"]

[tool.ruff.lint.pydocstyle]
convention = "google"

# Black configuration
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "matplotlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "scipy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "gtsam.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torch.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "torchvision.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "lightning.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "PIL.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cv2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pykitti.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pypose.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "osmnx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "networkx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "geopy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "haversine.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "streetview.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "requests.*"
ignore_missing_imports = true
